% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scclrR-package.R
\docType{package}
\name{scclrR-package}
\alias{scclrR}
\alias{scclrR-package}
\title{scclrR: Sparse PFlog normalization and PCA for Seurat}
\description{
\code{scclrR} stores PFlog as a sparse shifted-log matrix plus a per-cell
centering vector, and runs PCA on the implicit centered matrix through Rust.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/cleartools/scclrR}
  \item Report bugs at \url{https://github.com/cleartools/scclrR/issues}
}

}
\author{
\strong{Maintainer}: Lior Pachter \email{lpachter@caltech.edu}

}
\keyword{internal}
